Cleaning up source code.

+++ src/main/java/mod/akrivus/amalgam/client/render/layers/LayerBubbledItem.java
@@ -34,9 +34,9 @@ public class LayerBubbledItem implements LayerRenderer<EntityBubble> {
 			GlStateManager.popMatrix();
 		}
 		int color = bubble.getColor();
-		float r = (float)((color & 16711680) >> 16) / 255f;
-        float g = (float)((color & 65280) >> 8) / 255f;
-        float b = (float)((color & 255) >> 0) / 255f;
+		float r = ((color & 16711680) >> 16) / 255f;
+        float g = ((color & 65280) >> 8) / 255f;
+        float b = ((color & 255) >> 0) / 255f;
         GlStateManager.enableBlend();
         GlStateManager.color(r, g, b, 0.3F);
         this.renderer.bindTexture(new ResourceLocation("amalgam:textures/entities/bubble.png"));
